requirements.txt
Django==4.2.1
openpyxl==3.1.2
django-import-export==3.2.0